refactor: :recycle: split source code for readability and maintainability
